Minimum viable code
Lots of small changes to get a MVP.

We're including all the necessary GenX code and then overwritting the relevant pieces with DOLPHYN functions.

I took this include-based approach because if we include("GenX.jl") and/or using GenX we'll create a module which only exports a few of the functions that we need.

We could create a new function which separates the GenX and other pieces, but this keeps all the choices about whether to use GenX or alternative functions in one place