Debug Truck model and select Zone feature
Main branch had several errors related to Zone selection and Truck modeling features that led to many errors with Example System "ThreeZone_Liquid".

I have reverted all the changes related to Zone Selection and the modifications made to the truck branch to ensure that Example System "ThreeZone_Liquid" is solved without errors.

In future, we need to rigorous testing to make sure any features added are fully compatible with the all the examples on the main branch.