Fix solver configuration process,  add example using Gurobi, swap all to HiGHS
Previously we were silently loading Gurobi, CPLEX, etc if they were installed but not loaded. This meant we didn't have to ask the user to change the solver manually, but led to unusual behaviour (like it never working the first time it's run but then being fine afterwards) and is probably bad practice.

No the user has to define the solver when they create the runfile. It means the "solver" setting is a bit redundant and perhaps misleading. We should decide if we want to keep it or try a different approach.