Update force_TDR_recluster
Changed force_TDR_recluster to delete the existing TDR folder instead of justing using a check, as some parts of the TDR clustering / loading were using the TDR inputs instead of the original inputs.

This is because the code expected there to be no TDR inputs, so the load functions automatically used the TDR inputs if they existed and TDR was active. This could potentially lead to bugs elsewhere as clustering will be based on clustered inputs.