Renamed TruckEnergy to TruckChargePower
Renamed variables, expressions and parameters related to phrase "TruckEnergy" with "TruckChargePower" that better reflects the physical meaning of these terms.

TruckChargePower (previously labeled as TruckEnergy) refers to the power requirement associated with filling up a truck and implicitly limits the rate of filling up of the truck.
If the truck is a gas --> TruckChargePower refers to compressor charging requirement
If the truck is a liquid H2 --> Truck ChargePower could refer to pumping requirement or even liquefaction capacity